% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonProbEst.R
\docType{data}
\name{sampleP}
\alias{sampleP}
\title{A probabilistic sample}
\format{
An object of class \code{data.frame} with 500 rows and 5 columns.
}
\usage{
sampleP
}
\description{
A dataset of 500 individuals extracted with simple random sampling from a simulated fictitious population of 50,000 individuals. Further details on the generation of the dataset can be found in Ferri-García and Rueda (2018). The variables present in the dataset are the following:
\itemize{
  \item education_primaria. A binary variable indicating if the highest academic level achieved by the individual is Primary Education.
  \item education_secundaria. A binary variable indicating if the highest academic level achieved by the individual is Secondary Education.
  \item education_terciaria. A binary variable indicating if the highest academic level achieved by the individual is Tertiary Education.
  \item age. A numeric variable, with values ranging from 18 to 100, indicating the age of the individual.
  \item sex. A binary variable indicating if the individual is a man.
}
}
\references{
Ferri-García, R., & Rueda, M. (2018). \emph{Efficiency of propensity score adjustment and calibration on the estimation from non-probabilistic online surveys}. SORT-Statistics and Operations Research Transactions, 1(2), 159-162.
}
\keyword{datasets}
